requirements.txt
selenium
tqdm
reportlab
python-nmap
requests
beautifulsoup4
colorama
fuzzywuzzy
python-Levenshtein
cryptography
progress
python-dotenv
tensorflow
keras
scikit-learn
numpy
pandas
boto3
google-cloud-storage
azure-storage-blob
kubernetes
graphql-core
requests-toolbelt
matplotlib
seaborn
plotly
dnspython
lxml
html5lib
chardet
playwright
uvloop
aiohttp
jinja2
weasyprint
joblib
h5py
openapi-spec-validator
dotenv